done create todos, working on validate due_date
