+++ server/app.js
@@ -3,7 +3,6 @@ const app = express()
 const PORT = 3000
 const routes = require('./routes')
 
-app.use(express.JSON())
 app.use(express.urlencoded({ extended: true }))
 app.use(routes)
