--- server/controllers/controller.js
@@ -1,17 +1,0 @@
-const { Todo } = require('./models')
-
-class Controller {
-  static addTodo (req, res) {}
-
-  static showAllTodo (req, res) {}
-
-  static showTodoByID (req, res) {}
-
-  static editTodo (req, res) {}
-
-  static doneTodo (req, res) {}
-
-  static deleteTodo (req, res) {}
-}
-
-module.exports = Controller;
