+++ server/package.json
@@ -4,7 +4,8 @@
   "description": "Membuat website untuk mencatat hal - hal menarik untuk dilakukan",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "nodemon app.js"
   },
   "repository": {
     "type": "git",
