+++ server/routes/index.js
@@ -1,13 +1,13 @@
 const routes = require('express').Router()
-const Controller = require('../controllers/controller')
+const Controller = require('../controllers/todoController')
 
-routes.get('/', (req, res) => {
-  res.send('Welcome to another hell, enjoy your ride ^_^')
-})
+// routes.get('/', (req, res) => {
+//   res.send('Welcome to another hell, enjoy your ride ^_^')
+// })
 
 routes.post('/todos', Controller.addTodo)
 routes.get('/todos', Controller.showAllTodo)
-routes.get('/rodos/:id', Controller.showTodoByID)
+routes.get('/todos/:id', Controller.showTodoByID)
 routes.put('/todos/:id', Controller.editTodo)
 routes.patch('/todos/:id', Controller.doneTodo)
 routes.delete('/todos/:id', Controller.deleteTodo)
